Refactor chat components
