+++ src/components/Messages/ConversationList.tsx
@@ -22,12 +22,15 @@ const styles = {
 };
 
 const ConversationList = () => {
-  const { conversations, selectedConversation } = useAppSelector(
-    (state) => state.messages,
-  );
   const [messageModal, showMessageModal] = useState(false);
+
+  const conversations = useAppSelector((state) => state.messages.conversations);
+  const selectedConversationUserId = useAppSelector(
+    (state) => state.messages.selectedConversation.userId,
+  );
   const userId = useAppSelector((state) => state.user.userId);
   const dispatch = useAppDispatch();
+
   const navigate = useNavigate();
   const { sendRequest } = useAxios();
 
@@ -46,7 +49,7 @@ const ConversationList = () => {
   }, [dispatch, userId, sendRequest]);
 
   return (
-    <Box>
+    <>
       <Box sx={styles.header}>
         <Typography variant="h2">Messages</Typography>
         <IconButton onClick={() => showMessageModal(true)}>
@@ -65,7 +68,7 @@ const ConversationList = () => {
             onClick={() => {
               navigate(`/messages/${userId}/${o.otherUserId}`);
             }}
-            selected={selectedConversation.userId === o.otherUserId}
+            selected={selectedConversationUserId === o.otherUserId}
           />
         ))}
       </List>
@@ -73,7 +76,7 @@ const ConversationList = () => {
         onClose={() => showMessageModal(false)}
         open={messageModal}
       />
-    </Box>
+    </>
   );
 };
